pyproject.toml
[project]
name = "sco1-bumper"
version = "0.1.0"
description = "Automatically increment the project's version number."
authors = [
    {name = "sco1", email = "sco1.git@gmail.com"}
]
maintainers = [
    {name = "sco1", email = "sco1.git@gmail.com"}
]

readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

requires-python = ">=3.11"
dependencies = [
    "packaging~=24.1",
    "typer-slim~=0.12",
]

[project.urls]
Homepage = "https://github.com/sco1/"
Documentation = "https://github.com/sco1/bumper/blob/main/README.md"
Repository = "https://github.com/sco1/bumper"
Issues = "https://github.com/sco1/bumper/issues"
Changelog = "https://github.com/sco1/bumper/blob/main/CHANGELOG.md"

[project.scripts]
bumper = "bumper.cli:bumper_cli"

[tool.uv]
dev-dependencies = [
    "black~=24.10",
    "cogapp~=3.4",
    "flake8~=7.1",
    "flake8-annotations~=3.1",
    "isort~=5.13",
    "mypy~=1.11",
    "pre-commit~=4.0",
    "pytest~=8.3",
    "pytest-check~=2.4",
    "pytest-cov~=5.0",
    "pytest-randomly~=3.15",
    "ruff~=0.6",
    "tox~=4.18",
    "tox-uv~=1.11",
]

[tool.black]
line-length = 100

[tool.isort]
case_sensitive = true
known_first_party = "bumper,tests"
no_lines_before = "LOCALFOLDER"
order_by_type = false
profile = "black"
line_length = 100

[tool.mypy]
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.hatch.build.targets.wheel]
packages = ["bumper/"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
